+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - BUGFIX file extraction may fail for archives with directory entries without the D attribute [Issue 50](https://github.com/webdriverextensions/webdriverextensions-maven-plugin/issues/50). Thanks to [@ikucheryavenkov](https://github.com/ikucheryavenkov)
 - IMPROVEMENT Added user properties for most of the settings
 - IMPROVEMENT Support installing drivers when `skipTests` is true [Issue #49](https://github.com/webdriverextensions/webdriverextensions-maven-plugin/issues/49)
+- IMPROVEMENT add option to automatically set Selenium system properties [Issue #26](https://github.com/webdriverextensions/webdriverextensions-maven-plugin/issues/26)
 
 ## 3.2.0 (2019 June 2)
 - IMPROVEMENT Capability for custom file name with binaries [PR 42](https://github.com/webdriverextensions/webdriverextensions-maven-plugin/issues/42)